session table, add refresh token, api refresh toke

+++ db/sqlc/store.go
@@ -68,7 +68,7 @@ func (store *SQLStore) TransferTx(ctx context.Context, arg TransferTxParams) (Tr
 		result.Transfer, err = q.CreateTransfer(ctx, CreateTransferParams{
 			FromAccountID: arg.FromAccountID,
 			ToAccountID:   arg.ToAccountID,
-			Amount:        arg.Amount,
+			Amount:        int32(arg.Amount),
 		})
 
 		if err != nil {
