+++ db/sqlc/tranfers_test.go
@@ -13,7 +13,7 @@ func createRandomTransfer(t *testing.T, account1, account2 Account) Tranfer {
 	arg := CreateTransferParams{
 		FromAccountID: account1.ID,
 		ToAccountID:   account2.ID,
-		Amount:        util.RandMoney(),
+		Amount:        int32(util.RandMoney()),
 	}
 
 	transfer, err := testQueries.CreateTransfer(context.Background(), arg)
